+++ CHANGES.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0]
+* moved plotting functionality into dedicated modules.
+* made plotting dependencies check at plotting-function import, this removes the hard dependencies on matplotlib and corner.
+
 ## [1.2.0]
 * removed explicit print statements from warning catch and added suppress_warnings flag to suppress warnings*